# cmd/api/docs/swagger.yaml
definitions:
...
  models.InventoryUpdate:
    properties:
      product_id:
        type: integer
      stock:
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/inventories/add:
    post:
      consumes:
      - multipart/form-data
      description: Admin can add new  products
      parameters:
      - description: category_id
        in: formData
        name: category_id
        required: true
        type: string
      - description: product_name
        in: formData
        name: product_name
        required: true
        type: string
      - description: description
        in: formData
        name: description
        required: true
        type: string
      - description: price
        in: formData
        name: price
        required: true
        type: string
      - description: stock
        in: formData
        name: stock
        required: true
        type: string
      - description: image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Inventory
      tags:
      - Admin
  /admin/inventories/delete:
    delete:
      consumes:
      - application/json
      description: Admin can delete a product
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete Inventory
      tags:
      - Admin
  /admin/inventories/update:
    put:
      consumes:
      - application/json
      description: Admin can update stock of the inventories
      parameters:
      - description: update stock
        in: body
        name: add-stock
        required: true
        schema:
          $ref: '#/definitions/models.InventoryUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update Stock
      tags:
      - Admin
...
  /admin/products:
    get:
      consumes:
      - application/json
      description: client can view the list of available products
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Products
      tags:
      - Products
...
  /products:
    get:
      consumes:
      - application/json
      description: client can view the list of available products
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Products
      tags:
      - Products
  /products/category:
    get:
      consumes:
      - application/json
      description: client can filter with a category and get the list of  products
        in the category
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: category ID
        in: query
        name: catID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: filter Products by category
      tags:
      - Products
  /products/details:
    get:
      consumes:
      - application/json
      description: client can view the details of the product
      parameters:
      - description: Inventory ID
        in: query
        name: inventoryID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Show Product Details
      tags:
      - Products
  /products/search:
    get:
      consumes:
      - application/json
      description: client can search with a key and get the list of  products similar
        to that key
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: searchkey
        in: query
        name: searchkey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Search Products
      tags:
      - Products
...
swagger: "2.0"